backend/requirements.txt
fastapi
uvicorn
sqlmodel
passlib[bcrypt]
python-jose[cryptography]
python-multipart
paramiko
aiofiles
google-generativeai
python-dotenv
